winemac: Fix a string format warning for tick counts in update_clipboard().

+++ dlls/winemac.drv/clipboard.c
@@ -1783,7 +1783,7 @@ static void update_clipboard(void)
     static BOOL updating;
 
     TRACE("is_clipboard_owner %d last_clipboard_update %llu now %llu\n",
-          is_clipboard_owner, last_clipboard_update, GetTickCount64());
+          is_clipboard_owner, (unsigned long long)last_clipboard_update, (unsigned long long)GetTickCount64());
 
     if (updating) return;
     updating = TRUE;
